+++ pyproject.toml
@@ -74,9 +74,9 @@ SPS30     = "pypms.senserion.sps30"
 MCU680    = "pypms.bosch_sensortec.mcu680"
 
 [tool.poetry.plugins."pypms.extras"]
-influxdb = "pms.extra.influxdb:cli"
-mqtt     = "pms.extra.mqtt:cli"
-bridge   = "pms.extra.bridge:cli"
+influxdb = "pypms.extra.influxdb:cli"
+mqtt     = "pypms.extra.mqtt:cli"
+bridge   = "pypms.extra.bridge:cli"
 
 [tool.tox]
 legacy_tox_ini = """
